Move most logic to viewmodel
